requirement.txt
autopep8==1.6.0
click==8.1.2
cycler==0.11.0
Flask==2.1.1
fonttools==4.32.0
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
protobuf==3.20.0
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
typing-extensions==4.2.0
Werkzeug==2.1.1
zipp==3.8.0
